tests/wpt/tests/html/editing/dnd/drop/events-contenteditable-same-element-manual.tentative.html

<!doctype html>
<meta charset=utf-8>
<title>Selection drag and drop: events for contenteditable (same element)</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<body data-expected-events="
  b:drop:,
  b:beforeinput:deleteByDrag,
  b:input:deleteByDrag,
  b:beforeinput:insertFromDrop,
  b:textInput:,
  b:input:insertFromDrop">
<div contenteditable=true id=b data-select="0,7">Drag me ...to here:</div>
<script src="/uievents/textInput/support/common.js"></script>
<script src="support/events.js"></script>
